+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file.
 - #1622 Slow scrolling (search results, advanced settings and others)
 - #1626 AS3 decompilation - unpopped obfuscated function
 - #1624 Saving last searches saves only first results
+- #1627 Previously decompiled scripts not cached
+- SWF is not garbage collected on close in some situations
+- AS1/2 script search does not show all results
 
 ## [14.0.1] - 2021-02-26
 ### Added